add LED tester
This is similar to your led finder but it loops through each LED spits out the Tree LED Number, Broadcom GPIO Number, and Raspberry Pi Pin number. It uses ascii escape sequences to colour the results, green for on, red for off which makes for some nice christmas colours on the screen as well. If you feel it's redundant we can consolidate the two if needed.